apps/sector_classification_model_train/lambda/requirements.txt
sagemaker
sagemaker-experiments
boto3
python-dotenv